refactor: optimise media tables
