+++ lib/media_server/media_continues.ex
@@ -1,4 +1,4 @@
-defmodule MediaServer.Continues do
+defmodule MediaServer.MediaContinues do
   use Ecto.Schema
   import Ecto.Changeset
 
@@ -11,28 +11,25 @@ defmodule MediaServer.Continues do
     field :duration, :integer
 
     belongs_to :user, MediaServer.Accounts.User
-    belongs_to :media_type, MediaServer.MediaTypes
 
     timestamps()
   end
 
   def changeset(continue, attrs) do
     continue
-    |> cast(attrs, [:media_id, :current_time, :duration, :user_id, :media_type_id])
-    |> validate_required([:media_id, :current_time, :duration, :user_id, :media_type_id])
+    |> cast(attrs, [:media_id, :current_time, :duration, :user_id])
+    |> validate_required([:media_id, :current_time, :duration, :user_id])
   end
 
   def insert_or_update(attrs) do
     case Repo.get_by(__MODULE__,
            media_id: attrs.media_id,
            user_id: attrs.user_id,
-           media_type_id: attrs.media_type_id
          ) do
       nil ->
         %__MODULE__{
           media_id: attrs.media_id,
           user_id: attrs.user_id,
-          media_type_id: attrs.media_type_id
         }
 
       item ->
