--- lib/media_server/media_types.ex
@@ -1,33 +1,0 @@
-defmodule MediaServer.MediaTypes do
-  use Ecto.Schema
-  import Ecto.Changeset
-
-  alias MediaServer.Repo
-
-  schema "types" do
-    field :label, :string
-  end
-
-  def changeset(media_types, attrs) do
-    media_types
-    |> cast(attrs, [:type])
-    |> validate_required([:type])
-    |> unique_constraint(:type)
-  end
-
-  def get_movie_id() do
-    Repo.get_by!(__MODULE__, type: "movie").id
-  end
-
-  def get_series_id() do
-    Repo.get_by!(__MODULE__, type: "series").id
-  end
-
-  def get_episode_id() do
-    Repo.get_by!(__MODULE__, type: "episode").id
-  end
-
-  def get_type_id(type) do
-    Repo.get_by!(__MODULE__, type: type).id
-  end
-end
